Add default source functionality and tests (#107)
Completes the functionality previously hinted at with the default source function.  This is done by initializing the source setting when it does not exist.  We initialize it with an uninitialized source, which will automatically be initialized when it is opened.  If the source is removed or reset by name, the source list will be empty.  If the all sources are reset, thus deleting the source setting, then the next attempt to get the sources will result in reinitialization from the defaults.

Also adds another field to the source details, IsDefault.  This will give us options in the future to take action on these default sources, such as updating the URL if necessary.  We can write code to transform a source that matches on name and IsDefault being true without any impact to the users.