gs (#111)
Nothing is currently cleaning up installers from the temp location.  This change causes successful installs to remove the installer file, which should reduce the amount of data we leave around significantly.  When we support local installers, we will need to ensure that we don't attempt to delete them here, but for now since anything that we download is temporary, this should be ok.