Fix CTRL-C causing process exit due to code also using E_ABORT (#108)
Due to the way that CTRL signals were being handled, the code that would exit the process in response to multiple CTRL signals was being invoked if the cooperative cancellation was resulting in another E_ABORT context termination.  This change uses a unique error code for CTRL signals, thus removing the possibility of accidental early termination.