Change winget version determination and show OS and package info in --info (#115)
To make the version information that we display more consistent with our GitHub releases, this change updates the major & minor build numbers to be based solely on the version.h file that is used to set the actual binary version.  This is automatically updated for the GitHub builds by the script, but will need to be kept in sync to enable direct source pulls for now.  When inside a package, we do pull its build version as our own, to enable forward progress on the version number to be maintained in that manner.

Also changes the build script to use a two part version number, to put the build into the third part.

Finally, 'winget --info' will now print out the OS and package information, to better enable users to give debugging information.