Work towards enforcing output localization (#177)
This change introduces several new types and eases use of localized string resources.  The eventual goal is to have compile time enforcement of output localization.  That isn't quite achieved here, but the groundwork is laid.  What remains is largely moving strings into resources and tagging things appropriately.  There will be some thought/effort required in the TableOutput helper, but beyond that I don't see much trouble.

The new types for localized strings allow for direct output of a resource (Resource::StringId), making the code flow more naturally.  Additionally, if a string needs to be loaded manually to hold onto it, the Resource::LocString type takes a StringId as well.

The Utility::LocInd(View/String) types allow a value to be tagged as localization independent.  This is true for many strings that come from external sources, or are simply decorations in the output.  It can be applied to a string that is already localized, or that has no localization.

The Execution::Reporter::OutputStream has a commented line that will make compilation fail for any type not defined as allowed above.  This can be used to find additional code in need of update, and will eventually be enabled to enforce things long term.

Finally, several strings have been moved to resources, and many places have been changed for this format.  Many more remain for a future PR.  For now manual review should hopefully keep things from backsliding too much.