Change "Spring" to "March-May" to be more explicit (#354)
First of all, apologies if this should be an issue first, but would rather open a pr to show what I mean. 

A lot of roadmaps and timelines use seasons as targets which I personally really dislike. Seasons aren't the same throughout the world and so it can be a bit confusing for people in the southern hemisphere (like myself) and we normally just have to assume northern.

This may seem super minor and petty but I do think that it should honestly just be standard to go by either actual months, or even better quarters. (I prefer quarters but that is slightly different to seasons.) I also think we should be more exclusive in general of people throughout the world. 

Anyways, keen to make this sound better but I do really think this should be standard and am keen to here what everything thinks and hope people take it seriously. :)