Secure settings and sources (#455)
This change adds a secure setting type, and moves the source configuration to that type.  This also has the effect of leaving the old source configuration behind.  As multiple sources are not yet well supported, this is deemed an acceptable data loss.

The secure settings are secured by storing a hash of the stream on write.  This hash is stored in a location that requires administrator privileges to write, while the primary setting stream is stored in the app data.  If the app is uninstalled then reinstalled, or reset from settings, this stream will be lost and the sources reset.  The secure data will be left behind, but not referenced when no stream is present.  If a stream is present, its hash must also be present or an exception is thrown.  For sources, an attempt to tamper with the configuration would result in an unusable winget, until 'winget source reset --force' was executed, fully resetting the source data.

To support not writing to the source configuration, the sources are now layered.  This means that all user sources are read, then any default sources are added to the list.  To allow a default source to be removed, a tombstone concept is added that will allow the 'higher' layers to override the lower ones.

The update metadata is also separated out from the identity data, allowing it to be written without elevation.