Case insensitivity improvements (#459)
Makes version and channel case insensitive for lookups from the user.  They already were required to be case insensitive unique, so this shouldn't have any effect other than making use easier.

Causes an index modification (add / update) to overwrite the existing Id value with the incoming one if the casing is different.  So last writer wins on Id casing.  Other fields can have distinct casing per manifest revision for now, but since Id needs to be the thing that ties them together, it needs to be a single value.