Replace yaml-cpp with libyaml (#535)
Due to certain bugs in yaml-cpp, it has not had significant fuzzing applied.  Even if those bugs were fixed, it is unknown how many more might exist behind it.  In order to avoid chasing those bugs, we are switching to a YAML parser that does have significant fuzz testing coverage, [libyaml](https://github.com/yaml/libyaml).

In addition to the use of libyaml, a C++ wrapper is written to create a near identical surface area for use by the existing manifest validation.  It also includes encoding detection to enable support for ANSI (assumed Windows-1252, which wasn't supported by yaml-cpp), UTF-8 (with and without BOM), UTF-16 LE/BE (both with and without BOM).

Finally, the `validate` command now returns a non-zero value when there are issues.  If there is a failure during validation, the value returned is `0x8A150029`.  If there are only warnings, the value returned is `0x8A150028`.