Add fuzzing project for YAML manifest loading (#547)
This change adds a fuzzing project for YAML manifest loading, with the goal of exercising our code that sits atop libyaml (as well as libyaml).  It uses [libFuzzer](http://llvm.org/docs/LibFuzzer.html) to do this, which enables instrumented, coverage driven fuzzing.

Outside of the relatively simple fuzzing project, the major change is moving the manifest parsing logic from Repository to Common.  In addition, many small changes were required to make clang happy with our C++.  The biggest of these small changes is that all WinRT use is removed when building the fuzzing version of the lib.

Finally, both the WIL and C++/WinRT NuGet packages across all projects were updated to the latest available.