Add system reference values to manifest parsing (#574)
Adds the PackageFamilyName and ProductCode values to parsing by the manifest.  These can be at the root level, where they will be inherited.  They can also be provided per installer, but they must match the InstallerType or it will be an error.

Also makes major changes to the way the ManifestVersion is handled.  For now, the version is largely ignored (except it must be major version 0 if provided).  A concept of extensions is added, which are defined as the hyphen separated values after the primary version.  These can also be versioned, although the code does not exist to check it.  The syntax is:
```
ManifestVersion := Version (- Extension)
Extension := Version (- Extension)
```
where the Version portion of Extension is treated as `Name.Version`.  For instance, `1.0.0-extension` and `1.0.0-extension.2`.

Additionally:
- Fixes an issue with the version sorting where additional version characters would sort higher than those without (ex. 1.0 will now be greater than 1.0-alpha)