Refactor source and search functionality (#575)
Starting with `ISource`, a unique identifier is added to enable tying installs to them regardless of the name given to the source by the user.  `Search` is not changed at face value, but the result struct now returns a collection of `IPackage`s rather than the previous `IApplication`.  `IPackage` is a collection of `IPackageVersion`s in different states.  Currently that is [0, 1] possible installed versions, and [0, N] available versions.

`IPackageVersion` contains functionality to get individual properties of the specific version, as well as the manifest.

All other changes fall out of the above.