List command initial implementation behind feature toggle (#598)
This phase introduces the `list` command and much of the associative plumbing.  It currently only handles enumerating MSIX packages from the system, but it should be able to support additional enumeration simply by adding more to the index.

The goal of the `list` command is to enumerate packages installed on the machine, as well as to indicate which packages have an update available from a configured source.  This is achieved by creating predefined sources for installed packages, and a `CompositeSource` class to do the work of correlating the installed packages with those from external sources.