Implement manifest metadata in index (#626)
In order to store arbitrary amounts of data about installed packages, we need a more flexible table.  It isn't needed for available packages (yet), and so is only created on first write.  The data that is stored is { manifest rowid, metadata enum, value string }.  This part is very straightforward.

In order for the installed package enumeration for `list` et al. to be able to populate this, I realized it needed to be an enum stored in the repository code (sorry @yao-msft).  So part of this change is moving everything over to that.  In addition, I fixed some issues while I was around the code changing to an enum.

Finally, I hooked up the current MSIX enumeration to the only metadata item currently in use, and the SQLiteIndex source to read the metadata rather than just return an empty result every time.