Add ARP (Add/Remove Programs) data to list (#633)
With this change, the `list` command will now show packages from the machine and user Add\Remove Programs list.  This is done by reading from the registry locations that drive that UI, for both scopes and all appropriate architectures.

New common code is added to enable enumeration of registry keys and reading of values.  This is then leveraged in a set of helper functions to enable the existing installed package source to read in the relevant data.  It populates the index, and includes metadata relevant to `upgrade` and `uninstall`.

Entries marked as `SystemComponent` are excluded, as are those that do not have a `DisplayName`.  If a version is not discernable, it is marked as "Unknown".  This will lead it to be considered for potential upgrades until we properly track everything that winget installs and can also associate with the system data.