Add normalized name and publisher to index and correlation telemetry (#767)
## Change
This change adds the normalized name and publisher to the index, as well as leveraging this to collect telemetry on how well this normalization strategy is doing to correlate a manifest with the ARP entry on the machine.

The normalized name and publisher tables are added to the index as 1:N entries due to the introduction of full localization of name and publisher into the new manifest schema.  These tables don't have any link between them, which would allow a combination of name and publisher from any localization to work.  This tradeoff of precision seems worthwhile over writing more complex code to handle two values at a time, and it also allows the potential for one of the fields to change while the other remains constant and consume less space overall.

Searching is done by sending in the observed values for name and publisher, and the normalization is done by the source before comparing with its internal store.  In this way any type of normalization can occur internal to the source without the workflow logic needing to be aware.

In order to collect telemetry, we first snapshot the set of packages in ARP before executing the install.  Afterward, we compare the set with the snapshot to determine which ones are net new.  We also search the source using the manifest used to install to determine if there are any matches.  Finally, we intersect these two sets of packages.

The telemetry contains information about the package that was being installed `{source, id, version, channel}`, information about the number of packages in each state above `{changed, matching, intersection}`, and information from the ARP entry (if any) that we have decided is the one that represents the main package `{name, version, publisher, language}`.  We will be able to leverage this telemetry to improve both our normalization code and the manifests in winget-pkgs to improve our ability to correlate installed packages with remote ones.