Settings and command line argument for specifying scope (#819)
This change adds settings and a command line argument to `install` to enable control over the scope of the package install.  Scope is the difference between a user or machine install of a package.

The settings allow both the preference and requirement of scope to be set.  A preference affects the sort order for selecting installers, while a requirement filters the installers.  The default behavior (with no settings configured) is equivalent to:
```json
    "installBehavior": {
        "preferences": {
            "scope": "user"
        }
    },
```

The command line argument is `--scope` and accepts the values `user` and `machine`.  This is treated as a requirement, overriding any requirement set in the settings file.

The default value for Scope in a manifest is being changed from implicitly `User` to just `Unknown`.  If a requirement is specified (which cannot be `Unknown`), it will not match the default value of `Unknown`.  This is necessary to ensure that we can meet the requirement.  It will also affect what is a valid manifest; an unknown scope value will now conflict with any other scope value to prevent specifying an unknown and known value together.

Finally, to implement the actual behavior change, `ManifestComparator` is completely rewritten to be a more dynamic configuration of filters and comparators.  This will allow for more behavior to be added in the future in a straightforward manner.  It also fixes a bug that previously existed where the sorting was not consistent (ie. would not produce a properly sorted list).  It will now only use a lower priority sort when the higher priority sort considers the two installers equal.