Implement Delivery Optimization (#909)
## Change
This change adds code to leverage the Delivery Optimization (DO) service to download packages.  Only packages are downloaded through DO at this time, as other things that we download do not fit in with the design of DO.

Currently one must still opt-in to using DO via settings, which I used rather than an experimental feature because it makes sense to allow the control long term.  Eventually, we will make DO the default value here.

## Settings
Two settings are added:
1. Control over which downloader is used for packages (`.network.downloader`)
2. A timeout for controlling how long to wait before DO indicates any progress has been made (`.network.doProgressTimeoutInSeconds`)