Make additional metadata for downloads more extensible and add display name (#931)
In response to a comment after the PR was merged, update the Delivery Optimization download code to be able to set the DisplayName property.  We always set it to `Windows Package Manager`.

As part of this, I changed the way this metadata flows in to make it easier to add more to this optional data in the future.