Fix a few polish items (#966)
Fixes #723 and #728 , as well as an issue with what is displayed when multiple matches are found (add source) and anonymizes the error messages that can end up in telemetry.  Specifically, sometimes file paths are included in an STL exception message so we are replacing instances of the `%USERPROFILE%` path with that literal value.