Move Temp path to the user's location (#1146)
Move the temp directory that is used from the package specific temp to the user's temp location.  This should prevent the Mandatory Low ACL that was apparently causing #189 from being inherited.  While this is more avoiding the problem than fixing it, given the relatively low rate of occurrence it might be hard to ensure that it was truly fixed by actually changing the ACLs (and would require much more time and code to write).

Additionally changes the behavior of `Runtime::GetPathTo` to delete non-directories when they are encountered rather than just erroring directly.