Add deferred registration for MSIX (#1397)
Prior to this change, MSIX install (and upgrade) would forcibly shut down the application so that it could upgrade the package during the command.  After, the new package will be staged and the upgrade will complete when the application is restarted.

This is done by first attempting `RequestAddPackageAsync` if SmartScreen is necessary.  Then if it that fails due to the package being in use (or it is not needed at all), we stage (place files onto the machine) then register (install package for user).  If the package is in use at this point, we detect it and change the success message to indicate that the user must restart the application:
```
Successfully installed. Restart the application to complete the upgrade.
```