Add version comparison and architecture control to COM API (#1395)
This change adds an API, `CompareToVersion`, to compare a version string against a `PackageVersionInfo`.  The result is a value that indicates how the `PackageVersionInfo` object's version relates to the given value.  Thus if you want to check if the installed version is at least some value:
```C++
if (packageVersionInfo.CompareToVersion(targetVersion) == CompareResult::Lesser)
{
// Time to upgrade
}
```

It also adds a property to `InstallOptions`; `AllowedArchitectures`.  This vector of `ProcessorArchitecture` values contains both the allowed installer architectures that may be chosen when installing, as well as the order in which to prioritize the choice.  The vector starts with the default values for the current system.  Duplicate instances are ignored after the first, so in addition to completely replacing the set, one can prepend values as preferred without needing to understand all of the fallback options.

This change also fixes the default scope for `InstallOptions` to be `Any` as it is for the command line.

Finally, a new well known source is added that will contain a controlled set of "framework" packages for future use.  The packages within this source are not yet stabilized, so taking a dependency on it is not supported yet.