Source resilience (#1470)
Improves handling of errors from one source when multiple are configured by capturing the errors and including them as part of the source open and search results.  Some commands will continue to fail if one of the sources has an error to prevent said error from affecting a change to the system.  The commands that still fail are `install`, `import`, and `upgrade --all|<specific>` and they now indicate which source failed.

This change also replaces the experimental `msstore` source with the more official but still green `msstore` source (previously called `storepreview`).  This source is now capable of providing basic search functionality against the entire Store catalog.  It does not provide any correlation for installed packages however.  In addition, the store source is being moved out of experimental.