Handle single unknown version from REST source (#1483)
If a REST source returns a single version that is `Unknown`, look up all versions when the manifest is requested.

The update is triggered by a `GetManifest` call that meets the criteria (version is `Unknown` and manifest is missing).  When we see that, and the version info in the package is consistent, we perform an optimized search that should return the versions and manifests.  Then we overwrite the data from the `PackageVersion` that called us and it returns the manifest.