Fix correlation issue (#1492)
Update the name of the `SupportCorrelation` field to better reflect the intention and correct the use to not allow available source searching when installed only semantics are being used.

Intentionally does allow available/all package searching to correlate against local packages to enable `import` to detect when packages are already present.  Also updates `import` to work better with the REST optimized search path.