Make rename retry more frequently for longer, then try making a hardlink (#1497)
We already had a single retry after a 250ms wait for our `rename` of the installer to make it shell actionable with the correct extension.  This change makes it try every 100ms for 500ms, then use `copy_file` with a hard link option as a fall back.

The hard link is better than a true copy, as it won't take up more space.  But it won't get cleaned up on a successful install either.

Long term we should put in some background cleanup to remove old installers that were abandoned due to failures.