Settings streams exchange semantics (#1534)
The goal of this change is to add exchange semantics to settings stream writes.  Currently they do not have any interprocess synchronization.  This does not lead to corrupted streams as the underlying systems prevent simultaneous writes, but it can lead to data loss if updates from one process are not merged with those from another.

Because holding any kind of lock would likely lead to contention amongst the processes, we use exchange semantics instead.  Writes are only permitted if the stream has not changed since it was last read.  This required moving to a more object oriented model for the settings streams, which then required updating the consumers of it to also maintain an object and respond to failed write attempts.