Use package tracking data to correlate in CompositeSource (#1671)
This change leverages the tracking data that is stored per source to correlate installed packages to available packages.  This will currently provide no additional benefit with the `winget` source, as we aren't doing the work yet to determine the best fit system artifact when none of our current correlation attempts succeed.  It will help the `msstore` source to correlate as that is currently lacking in support to search system references.