Added an option to select architecture to install. (#1666)
Adds an option, `--architecture`, to `winget install`. This option allows the user to (you guessed it) tell winget what architecture they want to install. If the architecture isn't available, winget does the same thing it would do if there wasn't a installer entry for the user's platform (No applicable installer error). The user is only allowed to provide architectures their system can run (if they ask for arm64 on x64, a `CommandException` is thrown).