Add a default user agent to REST source calls (#1839)
Provides a default user agent header in the http helper class that the REST source code uses.  It can be overridden by providing the user agent header in the headers argument.

The default format is:
```
winget-cli WindowsPackageManager/1.2.3 DesktopAppInstaller/10.11.12
```
where `1.2.3` is the client version as would be output by `winget -v` (without the `v` at the beginning) and `10.11.12` is the package version of the MSIX.

Also updates the WinINET downloader to use the same user agent string, but it is not overridable.