Act on elevation requirements in majority cases (#2126)
When `ElevationRequirement` is set to `elevationProhibited` is set and winget is run elevated, block the installation with an error.

When `elevatesSelf` or `elevationRequired` is set and winget is not run elevated, inform the user that an elevation prompt is expected.  Also in the case of `elevationRequired`, use the `runas` verb on ShellExecute to run the installer elevated.  This can be used on installers that will fail if they are not run elevated.