Prevent null characters from getting into SQLite (#2289)
In order to prevent null characters from entering SQLite at all, the string binding functions are updated to throw if an embedded null is present.  Additionally, the consistency check is updated to detect strings with embedded null characters.

In order to prevent attempting to send null characters into the database at all, the `NormalizedString` type used throughout the manifest is updated to also replace any null characters with spaces.  This will allow those packages to be listed now.