Rewrite ApplyACL to allow for SYSTEM to participate. Grant SYSTEM full access to Temp as that is where we install from. (#2370)
Add support for `SYSTEM` to the explicit ACLs, and give `SYSTEM` full control on the `Temp` directory.

`ApplyACL` and the data that it uses are basically re-written to support a more dynamic list of ACEs.  Also ensures that if the process is running as `SYSTEM`, it doesn't attempt to write the same ACE multiple times.