Config code (#2942)
Introduces a new component, `Microsoft.Management.Configuration`.  This is the API surface for controlling the configuration of the system for a user (with the intention that any user can run it, including `SYSTEM`).  There are [extensive details on the API in this document](/JohnMcPMS/winget-cli/blob/config-code/doc/specs/Configuration-COM-API.md), so I will not repeat them here.

The API is not fully implemented by this change, only the baseline functionality for reading configuration files and acting on them immediately.  None of the history functionality is present, nor any ability to serialize files.  However, this should be sufficient to create an initial, golden path implementation of user interactive surfaces.