Configuration commands (#3036)
Adds the configuration commands `configure`, `configure show`, `configure test`, and `configure validate` (although it doesn't actually implement `test` or `validate`).

`configure show` outputs the configuration file that it is given, along with information on the configuration units contained within.

`configure` will first show the configuration file, then prompt the user to both review it and accept the responsibility for its actions. `--accept-configuration-agreements` can pre-accept the warning and apply the configuration without prompting. As the configuration is being applied, the individual results will be reported.

All of the configuration actions are run in a separate process to isolate winget.exe from a potential crash. That process is run at the same integrity level as winget.exe though, allowing elevated (admin) configuration actions.

Additionally, there are several minor fixes to things that I saw while developing this.